cmd: reorg commands under backup, add parsing
