+++ internal/cmd/cmd.go
@@ -63,7 +63,6 @@ func Run() {
 	registerImportCmd(rootCmd)
 	registerValidateCmd(rootCmd)
 	registerBackupCmd(rootCmd)
-	registerRestoreCmd(rootCmd)
 
 	// Register shared commands.
 	commands.RegisterPermissionCmd(rootCmd)
